% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMCCB.R
\name{CalcularBeneficio}
\alias{CalcularBeneficio}
\title{Control de Manejo de Ciclo Completo en Borregos (CMCCB)}
\usage{
CalcularBeneficio()
}
\value{
Ganancia, Gastos_totales, Ventas totales.
}
\description{
Esta función calcula el beneficio obtenido en un ciclo de producción de borregos,
tomando en cuenta los datos de peso, consumo de alimento y otros costos relacionados.
}
